+++ JudoPayDotNet/Models/PaymentModel.cs
@@ -2,7 +2,8 @@
 using System.Collections.Generic;
 using System.Runtime.Serialization;
 using JudoPayDotNet.Enums;
-using Newtonsoft.Json.Linq;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 
 namespace JudoPayDotNet.Models
 {
@@ -21,122 +22,52 @@ namespace JudoPayDotNet.Models
         }
 
         /// <summary>
-        /// Gets or sets your consumer reference.
+        /// The merchant reference to anonymously and uniquely identify a consumer.   GUIDs are recommended.
         /// </summary>
-        /// <value>
-        /// Your consumer reference.
-        /// </value>
         [DataMember(EmitDefaultValue = false)]
         public string YourConsumerReference { get; set; }
 
         /// <summary>
-        /// Gets your payment reference.
+        /// The merchant reference to uniquely identify a transaction.
         /// </summary>
-        /// <value>
-        /// Your payment reference.
-        /// PLEASE NOTE!!!! there is a reflection call within JudoPayClient.cs that gets this property via a string call. update in both places
-        /// including  other model instances of yourPaymentReference ********************
-        /// </value>
         [DataMember(EmitDefaultValue = false)]
         public string YourPaymentReference { get; set; }
 
         /// <summary>
-        /// Gets or sets your payment meta data.
+        /// The merchant payment meta data.
         /// </summary>
-        /// <value>
-        /// Your payment meta data.
-        /// </value>
         [DataMember(EmitDefaultValue = false)]
-        public IDictionary<string, string> YourPaymentMetaData { get; set; }
+        public IDictionary<string, object> YourPaymentMetaData { get; set; }
 
         /// <summary>
-        /// Gets or sets the judo identifier.
+        /// The judo account identifier.
         /// </summary>
-        /// <value>
-        /// The judo identifier.
-        /// </value>
         [DataMember(EmitDefaultValue = false)]
         public string JudoId { get; set; }
 
         /// <summary>
-        /// Gets or sets the amount.
+        /// The transaction amount.
         /// </summary>
-        /// <value>
-        /// The amount.
-        /// </value>
         [DataMember(EmitDefaultValue = false)]
         public decimal Amount { get; set; }
 
         /// <summary>
-        /// Gets or sets the currency.
+        /// The ISO currency code of the transaction (e.g. GBP, EUR, USD)
         /// </summary>
-        /// <remarks>Valid values GBP, EUR or USD.</remarks>
-        /// <value>
-        /// The currency.
-        /// </value>
         [DataMember(EmitDefaultValue = false)]
-// ReSharper disable MemberCanBePrivate.Global
         public string Currency { get; set; }
-// ReSharper restore MemberCanBePrivate.Global
 
         /// <summary>
-        /// Gets or sets the partner service fee.
-        /// </summary>
-        /// <value>
-        /// The partner service fee.
-        /// </value>
-        [DataMember]
-// ReSharper disable UnusedAutoPropertyAccessor.Global
-        public decimal PartnerServiceFee { get; set; }
-// ReSharper restore UnusedAutoPropertyAccessor.Global
-
-        /// <summary>
-        /// Gets or sets the consumer location.
-        /// </summary>
-        /// <value>
-        /// The consumer location.
-        /// </value>
-        [DataMember(EmitDefaultValue = false)]
-// ReSharper disable UnusedAutoPropertyAccessor.Global
-        public ConsumerLocationModel ConsumerLocation { get; set; }
-        // ReSharper restore UnusedAutoPropertyAccessor.Global
-
-        /// <summary>
-        /// This is a set of fraud signals sent by the mobile SDKs
+        /// Encrypted mobile device information sent from the mobile SDKs
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-// ReSharper disable once UnusedMember.Global
-        public JObject ClientDetails { get; set; }
+        public ClientDetailsModel ClientDetails { get; set; }
 
         /// <summary>
         /// Details needed for MCC 6012 transactions
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-// ReSharper disable once UnusedMember.Global
         public PrimaryAccountDetailsModel PrimaryAccountDetails { get; set; }
-
-        /// <summary>
-        /// The end consumers browser useragent for 3D secure
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-// ReSharper disable UnusedAutoPropertyAccessor.Global
-        public string UserAgent { get; set; }
-// ReSharper restore UnusedAutoPropertyAccessor.Global
-
-        /// <summary>
-        /// The end consumers browser DeviceCategory for 3D secure
-        /// </summary>
-        /// <remarks>This should either be Desktop or Mobile</remarks>
-        [DataMember(EmitDefaultValue = false)]
-        public string DeviceCategory { get; set; }
-
-        /// <summary>
-		/// The end consumers browser accept headers, used for 3D secure
-        /// </summary>
-        [DataMember(EmitDefaultValue = false)]
-// ReSharper disable UnusedAutoPropertyAccessor.Global
-        public string AcceptHeaders { get; set; }
-// ReSharper restore UnusedAutoPropertyAccessor.Global
 
         /// <summary>
         /// Is this transaction the first transaction of a series (has continuous authority
@@ -164,13 +95,21 @@ namespace JudoPayDotNet.Models
         /// Enum for Regular Payments (recurring)
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
+        [JsonConverter(typeof(StringEnumConverter))]
         public RecurringPaymentType? RecurringPaymentType { get; set; }
 
         /// <summary>
-        /// Receipt ID of original authenticated transaction (for recurring payments)
+        /// Judopay receipt ID of original authenticated transaction (for recurring or merchant initiated payments)
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
         public string RelatedReceiptId { get; set; }
+
+        /// <summary>
+        /// Payment network (scheme) transaction ID of original authenticated transaction (for recurring  or
+        /// merchant initiated payments)
+        /// </summary>
+        [DataMember(EmitDefaultValue = false)]
+        public string RelatedPaymentNetworkTransactionId { get; set; }
 
         /// <summary>
         /// Reference of associated web payment session