+++ JudoPayDotNet/Models/RegisterCardModel.cs
@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Runtime.Serialization;
 using JudoPayDotNet.Enums;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Converters;
 using Newtonsoft.Json.Linq;
 
 namespace JudoPayDotNet.Models
@@ -9,7 +11,6 @@ namespace JudoPayDotNet.Models
     /// Data to register a card (pre-auth with a pre-configured amount)
     /// </summary>
     [DataContract]
-    [KnownType(typeof(RegisterEncryptedCardModel))]
     // ReSharper disable UnusedMember.Global
     public class RegisterCardModel : SaveCardModel
     {
@@ -37,15 +38,6 @@ namespace JudoPayDotNet.Models
         /// </value>
         [DataMember(EmitDefaultValue = false)]
         public string YourPaymentReference { get; set; }
-
-        /// <summary>
-        /// Gets or sets the full card holder name.
-        /// </summary>
-        /// <value>
-        /// The card holder name.
-        /// </value>
-        [DataMember(IsRequired = false)]
-        public string CardHolderName { get; set; }
 
         /// <summary>
         /// Gets or sets the mobile number.
@@ -94,6 +86,7 @@ namespace JudoPayDotNet.Models
         /// Enum for Regular Payments (recurring)
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
+        [JsonConverter(typeof(StringEnumConverter))]
         public RecurringPaymentType? RecurringPaymentType { get; set; }
 
         /// <summary>